# server/internal/api/docs/swagger.yaml
basePath: /
definitions:
  api.ingestRequest:
    properties:
      device:
        type: string
      payload:
        type: object
      seq:
        type: integer
      type:
        type: string
    type: object
...
  store.Event:
    properties:
      created_at:
        type: string
      device:
        type: string
      id:
        type: integer
      payload:
        type: object
      seq:
        type: integer
      type:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Self-hosted reading and writing workspace. Syncs books, highlights,
    and reading positions across devices.
  title: Moringa API
  version: "1.0"
paths:
...
  /events:
    post:
      consumes:
      - application/json
      description: Accepts a sync event (highlight added, position updated, etc.)
        from a device, stores it in the event log, and broadcasts it to all connected
        clients via SSE.
      parameters:
      - description: Sync event payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.ingestRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/store.Event'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Ingest a sync event
      tags:
      - sync
...
  /stream:
    get:
      description: Opens a Server-Sent Events connection. Pass the current vector
        clock as query params (e.g. ?mac=44&phone=31). The server streams missed events
        first, then pushes new events in real time.
      parameters:
      - description: Last seen event seq for device 'mac'
        in: query
        name: mac
        type: integer
      - description: Last seen event seq for device 'phone'
        in: query
        name: phone
        type: integer
      produces:
      - text/event-stream
      responses:
        "200":
          description: OK
      security:
      - BearerAuth: []
      summary: SSE event stream
      tags:
      - sync
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"